+++ plugins/ucc_backup_restore/lib/ucc_backup_restore_web/views/admin_view.ex
@@ -4,6 +4,8 @@ defmodule UccBackupRestoreWeb.AdminView do
   """
   use UccBackupRestoreWeb, :view
 
+  import UcxUcc.Permissions
+
   @doc """
   The web url for the backup files.
   """