+++ plugins/ucc_chat/lib/ucc_chat_web/admin/layout.ex
@@ -12,7 +12,8 @@ defmodule UccChatWeb.Admin.Page.Layout do
       UccChatWeb.AdminView,
       "layout.html",
       75,
-      [pre_render_check: &UccChatWeb.Admin.view_message_admin_permission?/2]
+      pre_render_check: &check_perissions/2,
+      permission: "view-layout-administration"
     )
   end
 
@@ -23,4 +24,7 @@ defmodule UccChatWeb.Admin.Page.Layout do
     ], user, page, socket}
   end
 
+  def check_perissions(_page, user) do
+    has_permission? user, "view-layout-administration"
+  end
 end