+++ plugins/ucc_chat/lib/ucc_chat_web/admin/rooms.ex
@@ -4,13 +4,22 @@ defmodule UccChatWeb.Admin.Page.Rooms do
   import Ecto.Query
   import UcxUccWeb.Gettext
 
-  alias UcxUcc.Repo
+  alias UcxUcc.{Repo, Hooks}
   alias UccChat.Schema.Channel, as: ChannelSchema
 
   require Logger
 
   def add_page do
-    new("admin_rooms", __MODULE__, ~g(Rooms), UccChatWeb.AdminView, "rooms.html", 20)
+    new(
+      "admin_rooms",
+      __MODULE__,
+      ~g(Rooms),
+      UccChatWeb.AdminView,
+      "rooms.html",
+      20,
+      pre_render_check: &check_perissions/2,
+      permission: "view-room-administration"
+      )
   end
 
   def args(page, user, _sender, socket) do
@@ -18,9 +27,12 @@ defmodule UccChatWeb.Admin.Page.Rooms do
       preload: [:subscriptions, :messages])
 
     {[
-      user: user,
+      user: Repo.preload(user, Hooks.user_preload([])),
       rooms: rooms,
     ], user, page, socket}
   end
 
+  def check_perissions(_page, user) do
+    has_permission? user, "view-room-administration"
+  end
 end