+++ plugins/ucc_chat/lib/ucc_chat_web/admin/users.ex
@@ -10,7 +10,16 @@ defmodule UccChatWeb.Admin.Page.Users do
   # alias UccChat.{Message, Channel, UserService}
 
   def add_page do
-    new("admin_users", __MODULE__, ~g(Users), UccChatWeb.AdminView, "users.html", 30)
+    new(
+      "admin_users",
+      __MODULE__,
+      ~g(Users),
+      UccChatWeb.AdminView,
+      "users.html",
+      30,
+      pre_render_check: &check_perissions/2,
+      permission: "view-user-administration"
+    )
   end
 
   def args(page, user, _sender, socket) do
@@ -29,4 +38,7 @@ defmodule UccChatWeb.Admin.Page.Users do
     ], user, page, socket}
   end
 
+  def check_perissions(_page, user) do
+    has_permission? user, "view-user-administration"
+  end
 end