+++ plugins/ucc_chat/lib/ucc_chat_web/flex_bar/tab/room_info.ex
@@ -83,7 +83,7 @@ defmodule UccChatWeb.FlexBar.Tab.RoomInfo do
     end
   end
 
-  def flex_form_delete(socket, _sender, resource) do
+  def flex_form_delete(socket, _sender, _resource) do
     {:ok, socket}
   end