+++ plugins/ucc_webrtc/lib/ucc_webrtc_web/admin/webrtc.ex
@@ -7,7 +7,16 @@ defmodule UccWebrtcWeb.Admin.Page.Webrtc do
   alias UccWebrtc.Settings.Webrtc
 
   def add_page do
-    new("admin_webrtc", __MODULE__, ~g(WebRTC), UccWebrtcWeb.AdminView, "webrtc.html", 90)
+    new(
+      "admin_webrtc",
+      __MODULE__,
+      ~g(WebRTC),
+      UccWebrtcWeb.AdminView,
+      "webrtc.html",
+      90,
+      pre_render_check: &check_perissions/2,
+      permission: "view-webrtc-administration"
+    )
   end
 
   def args(page, user, _sender, socket) do
@@ -17,4 +26,7 @@ defmodule UccWebrtcWeb.Admin.Page.Webrtc do
     ], user, page, socket}
   end
 
+  def check_perissions(_page, user) do
+    has_permission? user, "view-webrtc-administration"
+  end
 end